refactor(024): move the shared ai core into modules/aicore with shims
Pure move of the provider factory and its three adapters, usage and
pricing, the instruction guard, the single model call (askModel and
parseModelJson, lifted out of the agent orchestrator) and the LangGraph
persistence factory into Modules/AICore. Every old path is a one-line
re-export shim so the consumers and their jest.mock calls keep working;
they are repointed one module per pull request in the follow-ups.

ProjectTemplates called api.openai.com directly with axios; it now goes
through getProvider().chat like every other model call, and a conventions
test keeps vendor SDKs and hosts inside Modules/AICore from here on.
