+++ Modules/Agents/engine/orchestrator.js
@@ -1,6 +1,5 @@
-const logger = require('../../../Config/loggerConfig');
-const { getProvider, isAnyProviderConfigured } = require('../../AIProjectGenerator/llmProvider');
-const { emptyUsage, usageFromResult, addUsage } = require('../../AIProjectGenerator/usage');
+const { emptyUsage } = require('../../AIProjectGenerator/usage');
+const { askModel, parseModelJson } = require('../../AICore/modelCall');
 const { audit, extractUrl } = require('./pageAudit');
 const { isBlockedHostname } = require('./safeFetch');
 const skillIndex = require('../skills');
@@ -17,7 +16,6 @@ const skillIndex = require('../skills');
 // output, and a hallucinated finding has nowhere to enter — `verify` drops
 // anything whose factId was not measured.
 
-const LOG_PREFIX = '[agent]';
 const SKILLS = skillIndex.BY_SLUG;
 const GATHERED = 'gathered';
 
@@ -31,14 +29,6 @@ const requireSkill = (slug) => {
 
 const SEVERITY = ['high', 'medium', 'low'];
 const rank = (s) => { const i = SEVERITY.indexOf(String(s).toLowerCase()); return i === -1 ? 2 : i; };
-
-/* Strip markdown fences some models wrap JSON in, then parse. A parse failure is
- * a failed run, never a silent empty result — silence would read as "clean page". */
-function parseModelJson(raw) {
-    const text = String(raw || '').trim().replace(/^```(?:json)?\s*/i, '').replace(/```$/, '').trim();
-    try { return { ok: true, value: JSON.parse(text) }; }
-    catch (e) { return { ok: false, error: `model did not return valid JSON: ${e.message}` }; }
-}
 
 /* PHASE 4 — verify. The gate that makes this trustworthy.
  *  - evidence: a finding must name a fact id that actually failed
@@ -88,35 +78,6 @@ function findingsWithoutModel(auditResult, skill) {
         fix: f.evidence ? `Evidence: ${f.evidence}` : '',
         evidence: f.detail,
     }));
-}
-
-/* The one model call. `raw` stays null when the provider is missing, fails or
- * answers with something that is not JSON; `degraded` says which. */
-async function askModel(skill, { prompt, budget }) {
-    let usage = emptyUsage();
-    let raw = null; let model = null; let degraded = null;
-    if (isAnyProviderConfigured() && budget.allowModel !== false) {
-        try {
-            const provider = getProvider();
-            const result = await provider.chat({
-                systemPrompt: skill.systemPrompt,
-                messages: [{ role: 'user', content: prompt }],
-                maxTokens: Math.min(skill.maxTokens, budget.maxTokens || skill.maxTokens),
-                temperature: 0.2,
-                jsonMode: true,
-            });
-            usage = addUsage(usage, usageFromResult(result));
-            model = result.model || null;
-            const parsed = parseModelJson(result.content);
-            if (parsed.ok) raw = parsed.value; else degraded = parsed.error;
-        } catch (error) {
-            degraded = `model call failed: ${error.message}`;
-            logger.error(`${LOG_PREFIX} ${degraded}`);
-        }
-    } else {
-        degraded = 'no LLM provider configured';
-    }
-    return { raw, model, degraded, usage };
 }
 
 const skipped = (skill, reason, started) => ({ status: 'skipped', reason, skill: skill.slug, findings: [], usage: emptyUsage(), durationMs: Date.now() - started });
